+++ packages/theme_tailor/lib/src/model/library_data.dart
@@ -1,0 +1,9 @@
+class ImportsData {
+  const ImportsData({
+    required this.hasJsonSerializable,
+    required this.hasFlutterDiagnosticable,
+  });
+
+  final bool hasJsonSerializable;
+  final bool hasFlutterDiagnosticable;
+}