+++ packages/theme_tailor_annotation/lib/src/theme_tailor.dart
@@ -1,5 +1,8 @@
+import 'package:json_annotation/json_annotation.dart';
 import 'package:meta/meta_meta.dart';
 import 'package:theme_tailor_annotation/theme_tailor_annotation.dart';
+
+part 'theme_tailor.g.dart';
 
 const tailor = Tailor();
 
@@ -8,22 +11,29 @@ const tailor = Tailor();
 ///
 /// {@endtemplate}
 @Target({TargetKind.classType})
+@JsonSerializable(
+  fieldRename: FieldRename.snake,
+  createToJson: false,
+  anyMap: true,
+)
 class Tailor {
   /// {@macro theme_tailor.theme_tailor}
   const Tailor({
     this.themes,
-    this.themeGetter = ThemeGetter.onBuildContextProps,
+    this.themeGetter,
     this.encoders,
-    this.requireStaticConst = false,
+    this.requireStaticConst,
     this.generateStaticGetters = true,
   });
+
+  factory Tailor.fromJson(Map json) => _$TailorFromJson(json);
 
   final List<String>? themes;
 
   /// Create getters for the easy access of the theme properties
   /// In case of creating component/modular themes, set it to
   /// [ThemeGetter.none]
-  final ThemeGetter themeGetter;
+  final ThemeGetter? themeGetter;
 
   /// A list of [ThemeEncoder]s to apply to this class.
   /// If this is null, default encoders will be used
@@ -54,6 +64,8 @@ class Tailor {
   /// @myCustomAnnotation
   /// class OtherExampleTheme {...}
   /// ```
+  // ignore: deprecated_member_use
+  @JsonKey(ignore: true)
   final List<ThemeEncoder>? encoders;
 
   /// If true, the generator will force generating constant themes.
@@ -74,7 +86,7 @@ class Tailor {
   ///   static const otherList = someOtherList;
   /// }
   /// ```
-  final bool requireStaticConst;
+  final bool? requireStaticConst;
 
   /// If true, static getters will be generated to support updating theme
   /// properties on hot reload. They will conditionally return either a getter